added -ts flag to vc.ml
This flag will allow us to just print scanned tokens from our
the inputed file.

The way I see this working is by using the following command:

	dune exec -- vc test/helloworld.viz -a > test.out

Now we can do a diff between test.out and some predefined test file
to ensure that we are scanning the code correctly.

We would then need to do this for ast, sast, ..., and the final
outputted program once we get there.