big commit
2 things:

in Header I added print functions, this print_token will take a parse
token and then print out what its value it. This is used for -ts
flag, and will be used for writing scanner tests.

in scanner body, I split ID scan into both function and variable
as we define in lrm. also added float.

for floats and ints, we will leverage ocaml builtins which will throw
an error if the number is too big.