committed the start of a comprehensive test suite
we need to follow naming conventions

filenames that start with fail imply errors and can only have .err
or .ref tags

filenames that start with test imply valid tests and can only have
.viz or .ref tags

now on to writing a test script that should produce a .out file using
the dune exec command with -ts flags. do a diff between the .out
and .ref file and print the output.

never commit .out files, and after the test script runs be sure to clean
the directory.