committing a test script
This test script could be run for all of the tests in the scanner
directory, you just need to follow the naming conventions.

.viz -> .out, and diff .out .ref
.err -> .out, and diff .out .ref

Now we have a more systematic/automated test suite. Further, this
test script can be repurposed for the other phases of testing
such as parser, ast, sast and more.

At a min, we can test our scanner more thoroughly.