support for more binary ops
==, !=, <, >, <=, >=

all of the changes have been propagated throughout the compiler files.
Also, to avoid shift/reduce errors I needed to add in precedence
amongst the token types. This is something that was missing in my
first rendition way back when. I added all of the token operators that
correspond to C precedence standard.

Also, formatted some of the print functions to print a little prettier
to allow me to do more tests for parsing and semantics.