added support for unary ops
Well, we only support one right now "Not" But the framework is in
play. just need to update the semantic checker which only looks
to see if the passed type is BoolType.

That is an easy fix. The other unary operators that I can see us
adding would be (typecast, ++ and -- for both post and prefix).