added if, while, return
the parser and AST, SAST will work.

semantics are not working, I think it is a limitation of our current
CFG. I will not continue, until we meet as a group and choose a concrete
plan.

we are at a point where i dont want work to be wasted. lets talk later