refactored vc.ml file
This is the dune build file which will perform a variety of actions.

either do scanning, parsing, semantic checking, or compile fully.

That is what the flags are for.

just do dune clean and dune build. or use the specific flags for
your tests.

when we do IR and codegen, we will need to add more flags.