working on logic.
we can now support the following logic flows

if () {...}
if () {...} else {...}
(expr) ? true stmt; : false stmt;

To facilitate this I needed to add a No Op where if there is no else
branch like above we create an AST for if(predicate, true block, No_op).

This doesnt break anything and there are no shift/reduce errors.

We just need to check if s2, in if(predicate, s1, s2) is a No_op when
doing code gen.

Further, look at test/parser/logic for examples. It passes the -a
and -s checks