very big commit
we now have functionality for our for loop constructs.

we wrote in LRM that we wanted:

for counter in starting_num ... <ending condition>ending_num step step_number

That is implemented. The code builds with no shift/reduce errors
and my test/programs/loops.viz file will pass all checks.

run the test/programs/loops.viz file with -a and -s flags to double
check my work.

I think that this is correct. Semantic file also compiles. I essentially
mapped our for loop readable syntax, to that for a normal For(...)
loop