handle the no_op
no_op is a place holder for an if without an else block.

So in the SAST if we see a SNo_op then we just replace it with an
empty SBlock([])

all of the loop constructs still pass the semantics check. And now,
there will be no issue with codegen