made the changes matt and yanhao discussed
removed no_op, and replaced it with NOELSE and %prec.

changed ... to RANGE token.

cannot change the ternary issue, because of the way if statement
is defined. it expects expr * stmt * stmt, not expr * ecpr * expr.

what I have to do is now add in a new semantic/ast/sast target, I
will do that later.