another hefty commit
what works:

added NoneLit which will be returned in cases where we have a function
return type of none.

now you can type `return;` for none functions, or not type anything
at all like you normally would in a void func. This works.

To support void return types, I needed to borrow some
code from MicroC.

But basically we need to add in a void instrcutions into a couple
of places in code gen.

all of the code compiles, and runs through the codegen phase

what doesnt work:

printing of different data types. I began the process of installing
different builtin print functions but these are not done. The test will fail
but of course the code compiles and runs