updated readme
please read it, for updates on how to build and run the code.

next, the print code almost works now. it is mapping the print to
the correct version we want (i.e. print_int, print_float, ...) but
instead of the expression being printed the format is being printed

%d, %B, ...

step in the right direction, it is close to working