fixed arithmetic
all of the arithmetic works except for float %= float. But
weirdly float % float works.

all arithmetic and assignment works!

I had to change the scanner, because there was a parser error on
1-1. It scanned it as int int, as opposed to int minus int. Thus,
tests are failing on negative numbers.

I will next be implementing a negate Unary op, so that we could do
x: int = -123;